--- packages/deployment-service/src/entities/index.ts
@@ -1,3 +1,0 @@
-export * from './pipeline.entity'
-export * from './pipeline-runner.entity'
-export * from './task.entity'
