+++ packages/deployment-service/src/functions/pipeline/pipeline-create.ts
@@ -1,7 +1,7 @@
 import { httpHandler, ValidationException } from '@homeservenow/serverless-aws-handler'
-import { PipelineDto } from '@/dto'
-import { PipelineEntity } from '@/entities'
-import * as service from '@/services/pipeline'
+import { PipelineDto } from './../../dto'
+import { PipelineEntity } from './../../entities/pipeline.entity'
+import * as service from './../../services/pipeline'
 import { plainToClass } from 'class-transformer'
 import { validate } from 'class-validator'
 import { resolveDeveloperId } from '../../utils'
