+++ packages/deployment-service/src/functions/pipeline/pipeline-runner-create.ts
@@ -1,5 +1,5 @@
-import { PipelineEntity } from '@/entities'
-import { ownership, resolveDeveloperId } from '@/utils'
+import { PipelineEntity } from './../../entities/pipeline.entity'
+import { ownership, resolveDeveloperId } from './../../utils'
 import { httpHandler, NotFoundException } from '@homeservenow/serverless-aws-handler'
 import * as service from '../../services'
 import { defaultOutputHeaders } from './../../constants'
