+++ packages/deployment-service/src/functions/pipeline/pipeline-runner-get.ts
@@ -1,5 +1,5 @@
-import { PipelineRunnerEntity } from '@/entities'
-import { ownership, resolveDeveloperId } from '@/utils'
+import { PipelineRunnerEntity } from './../../entities/pipeline-runner.entity'
+import { ownership, resolveDeveloperId } from './../../utils'
 import { httpHandler, NotFoundException } from '@homeservenow/serverless-aws-handler'
 import * as service from '../../services'
 import { defaultOutputHeaders } from './../../constants'
