+++ packages/deployment-service/src/functions/pipeline/pipeline-runner-update.ts
@@ -1,5 +1,5 @@
-import { PipelineEntity } from '@/entities'
-import { ownership, resolveDeveloperId } from '@/utils'
+import { PipelineEntity } from './../../entities/pipeline.entity'
+import { ownership, resolveDeveloperId } from './../../utils'
 import { httpHandler, BadRequestException, NotFoundException } from '@homeservenow/serverless-aws-handler'
 import { DeploymentStatus } from '@reapit/foundations-ts-definitions'
 import * as service from '../../services'
